Use canonical TF convention for ops/kernels
* Move op registration to ops/ngraph_encapsulate_op.cc
* Move kernel registration/implementation to kernels/ngraph_encapsulate_op.cc
* Remove ngraph_encapsulate_impl
* Remove unused ngraph_register_stub_kernels.{c,h}
* Remove unit test for ngraph_encapsulate_op for now
