+++ ngraph_bridge/kernels/ngraph_encapsulate_op.cc
@@ -33,8 +33,6 @@
 #include "ngraph_bridge/ngraph_backend_manager.h"
 #include "ngraph_bridge/ngraph_builder.h"
 #include "ngraph_bridge/ngraph_cluster_manager.h"
-#include "ngraph_bridge/ngraph_encapsulate_impl.h"
-#include "ngraph_bridge/ngraph_encapsulate_op.h"
 #include "ngraph_bridge/ngraph_mark_for_clustering.h"
 #include "ngraph_bridge/ngraph_timer.h"
 #include "ngraph_bridge/ngraph_utils.h"
@@ -44,35 +42,68 @@ using namespace std;
 namespace tensorflow {
 namespace ngraph_bridge {
 
-int NGraphEncapsulateOp::s_instance_id = 0;
+class NGraphEncapsulateOp : public OpKernel {
+ public:
+  explicit NGraphEncapsulateOp(OpKernelConstruction* ctx);
+  ~NGraphEncapsulateOp() override;
+  void Compute(OpKernelContext* ctx) override;
 
-//---------------------------------------------------------------------------
-//  NGraphEncapsulateOp::ctor
-//---------------------------------------------------------------------------
+ private:
+  Status GetNgExecutable(
+      const std::vector<Tensor>& tf_input_tensors,
+      std::shared_ptr<Executable>& ng_exec);
+
+  std::mutex m_compute_lock_;
+  Graph m_graph;
+  int m_cluster_id;
+  int m_function_cache_depth_in_items = 16;
+  string m_name;
+  std::vector<bool> m_input_is_static;
+  std::list<std::string> m_lru;
+  std::unordered_map<std::string, std::shared_ptr<Executable>> m_ng_exec_map;
+};
+
+static Status ParseNodeAttributes(
+    const google::protobuf::Map<string, AttrValue>& additional_attributes,
+    std::unordered_map<std::string, std::string>* additional_attribute_map) {
+  for (auto itx : additional_attributes) {
+    // Find the optional attributes to be sent to the backend.
+    // The optional attributes have '_ngraph_' appended to the start
+    // so we need to get rid of that and only send the remaining string
+    // since the backend will only look for that.
+    // '_ngraph_' is only appended for the bridge.
+    // For e.g. _ngraph_ice_cores --> ice_cores
+    if (itx.first.find("_ngraph_") != std::string::npos) {
+      // TODO: decide what the node attributes should be.
+      // right now _ngraph_ is used for optional attributes
+      auto attr_name = itx.first;
+      auto attr_value = itx.second.s();
+      NGRAPH_VLOG(4) << "Attribute: " << attr_name.substr(strlen("_ngraph_"))
+                     << " Value: " << attr_value;
+      additional_attribute_map->insert(
+          {attr_name.substr(strlen("_ngraph_")), attr_value});
+    }
+  }
+  return Status::OK();
+}
+
 NGraphEncapsulateOp::NGraphEncapsulateOp(OpKernelConstruction* ctx)
-    : OpKernel(ctx) {
+    : OpKernel(ctx), m_graph(OpRegistry::Global()) {
   NGRAPH_VLOG(1) << "Create Executor " << name();
-  ng_encap_impl_.SetName(name());
+  m_name = name();
 
+  OP_REQUIRES_OK(ctx, ctx->GetAttr<int>("ngraph_cluster", &m_cluster_id));
   std::ostringstream oss;
-  oss << "Encapsulate_" << ng_encap_impl_.GetInstanceId() << ": " << name();
-
+  oss << "Encapsulate_" << m_cluster_id << ": " << name();
   NG_TRACE(oss.str(), name(), "");
 
-  NGRAPH_VLOG(1) << "NGraphEncapsulateOp: " << ng_encap_impl_.GetInstanceId()
+  NGRAPH_VLOG(1) << "NGraphEncapsulateOp: " << m_cluster_id
                  << " Name: " << name();
 
-  GraphDef* graph_def;
-
-  int cluster{-1};
-  OP_REQUIRES_OK(ctx, ctx->GetAttr<int>("ngraph_cluster", &cluster));
-  ng_encap_impl_.SetNgraphCluster(cluster);
-  graph_def =
-      NGraphClusterManager::GetClusterGraph(ng_encap_impl_.GetNgraphCluster());
-
+  GraphDef* graph_def = NGraphClusterManager::GetClusterGraph(m_cluster_id);
   if (graph_def == nullptr) {
     string flib_key =
-        "ngraph_cluster_" + to_string(ng_encap_impl_.GetNgraphCluster());
+        "ngraph_cluster_" + to_string(m_cluster_id);
     // Read graphdef from function library
     const FunctionLibraryDefinition flib =
         *ctx->function_library()->GetFunctionLibraryDefinition();
@@ -91,17 +122,14 @@ NGraphEncapsulateOp::NGraphEncapsulateOp(OpKernelConstruction* ctx)
     if (!status.ok()) {
       NGRAPH_VLOG(2) << "FunctionDefToBodyHelper returned a not ok status.";
     }
-    CopyGraph(*fnbody->graph, &ng_encap_impl_.m_graph);
+    CopyGraph(*fnbody->graph, &m_graph);
   } else {
     GraphConstructorOptions opts;
     opts.allow_internal_ops = true;
     OP_REQUIRES_OK(
-        ctx, ConvertGraphDefToGraph(opts, *graph_def, &ng_encap_impl_.m_graph));
+        ctx, ConvertGraphDefToGraph(opts, *graph_def, &m_graph));
   }
 
-  int graph_id{-1};
-  OP_REQUIRES_OK(ctx, ctx->GetAttr("ngraph_graph_id", &graph_id));
-  ng_encap_impl_.SetGraphId(graph_id);
   //
   // Initialize the "m_input_is_static" vector as follows:
   // (1) create m_input_is_static with n+1 elements, where n is the max arg
@@ -114,7 +142,7 @@ NGraphEncapsulateOp::NGraphEncapsulateOp(OpKernelConstruction* ctx)
   int32 max_arg_index = -1;
   std::vector<const Node*> arg_nodes;
 
-  for (auto node : ng_encap_impl_.m_graph.nodes()) {
+  for (auto node : m_graph.nodes()) {
     if (node->type_string() == "_Arg") {
       arg_nodes.push_back(node);
 
@@ -125,13 +153,12 @@ NGraphEncapsulateOp::NGraphEncapsulateOp(OpKernelConstruction* ctx)
   }
 
   int size = max_arg_index + 1;
-  ng_encap_impl_.ResizeStaticInputVector(size);
+  m_input_is_static.resize(size);
 
   for (int i = 0; i < size; i++) {
-    ng_encap_impl_.SetStaticInputVector(i, false);
+    m_input_is_static[i] = false;
   }
 
-  // Fill the vector.
   for (auto node : arg_nodes) {
     int32 index;
     OP_REQUIRES_OK(ctx, GetNodeAttr(node->attrs(), "index", &index));
@@ -152,52 +179,42 @@ NGraphEncapsulateOp::NGraphEncapsulateOp(OpKernelConstruction* ctx)
       }
     }
     NGRAPH_VLOG(5) << "Marking arg " << index << " is_static: " << is_static;
-    ng_encap_impl_.SetStaticInputVector(index, is_static);
+    m_input_is_static[index] = is_static;
   }
 
   // Get the optional attributes
   std::unordered_map<std::string, std::string> additional_attribute_map;
   auto node_def = ctx->def();
-  OP_REQUIRES_OK(ctx, ng_encap_impl_.ParseNodeAttributes(
+  OP_REQUIRES_OK(ctx, ParseNodeAttributes(
                           node_def.attr(), &additional_attribute_map));
 }
 
-//---------------------------------------------------------------------------
-//  ~NGraphEncapsulateOp()
-//---------------------------------------------------------------------------
 NGraphEncapsulateOp::~NGraphEncapsulateOp() {
   std::ostringstream oss;
-  oss << "Destroy Encapsulate_" << ng_encap_impl_.GetInstanceId() << ": "
+  oss << "Destroy Encapsulate_" << m_cluster_id << ": "
       << name();
   NG_TRACE(oss.str(), name(), "");
   NGRAPH_VLOG(2) << "~NGraphEncapsulateOp::" << name();
-  ng_encap_impl_.ClearExecMaps();
+  m_ng_exec_map.clear();
 }
 
-//---------------------------------------------------------------------------
-// OpKernel::Compute
-//---------------------------------------------------------------------------
 void NGraphEncapsulateOp::Compute(OpKernelContext* ctx) {
   NGRAPH_VLOG(1) << "Compute using executor " << name();
   std::ostringstream oss;
-  oss << "Execute: Encapsulate_" << ng_encap_impl_.GetInstanceId() << ": "
+  oss << "Execute: Encapsulate_" << m_cluster_id << ": "
       << name();
   NG_TRACE(oss.str(), name(), "");
+  NGRAPH_VLOG(4) << "NGraphEncapsulateOp::Compute starting for cluster "
+                 << m_cluster_id;
 
   Timer compute_time;
   std::lock_guard<std::mutex> lock(m_compute_lock_);
-  int cluster_id = ng_encap_impl_.GetNgraphCluster();
-  NGRAPH_VLOG(4) << "NGraphEncapsulateOp::Compute starting for cluster "
-                 << cluster_id;
   int time_func_create_or_lookup;
   Timer function_lookup_or_create;
 
-  std::vector<TensorShape> input_shapes;
-  std::vector<const Tensor*> static_input_map;
-  std::shared_ptr<Executable> ng_exec;
-
   // TF input tensor
   std::vector<Tensor> tf_input_tensors;
+  std::shared_ptr<Executable> ng_exec;
   int step_id;
   {
     NG_TRACE("FunctionMaybeCreate", name(), "");
@@ -209,34 +226,44 @@ void NGraphEncapsulateOp::Compute(OpKernelContext* ctx) {
 
     // Get ngraph executable and inputs information
     OP_REQUIRES_OK(
-        ctx, ng_encap_impl_.GetNgExecutable(tf_input_tensors, input_shapes,
-                                            static_input_map, ng_exec));
+        ctx, GetNgExecutable(tf_input_tensors, ng_exec));
 
     NGRAPH_VLOG(1) << " Step_ID: " << step_id;
     NGRAPH_VLOG(4)
         << "NGraphEncapsulateOp::Compute got ngraph executable for cluster "
-        << cluster_id;
+        << m_cluster_id;
 
     time_func_create_or_lookup = function_lookup_or_create.ElapsedInMS();
   }
 
   NGRAPH_VLOG(4) << "NGraphEncapsulateOp::Compute got graph for cluster "
-                 << cluster_id;
+                 << m_cluster_id;
 
   Timer create_or_lookup_tensors;
-
-  // Allocate tensors for input arguments.
   vector<shared_ptr<ngraph::runtime::Tensor>> ng_inputs;
   int ng_input_tensor_size_in_bytes = 0;
   {
     NG_TRACE("Input: maybe create", name(), "");
-    OP_REQUIRES_OK(
-        ctx, ng_encap_impl_.AllocateNGTensors(tf_input_tensors, ng_inputs));
+    // Allocate tensors for input arguments.
+    for (int i = 0; i < tf_input_tensors.size(); i++) {
+      ngraph::Shape ng_shape(tf_input_tensors[i].shape().dims());
+      for (int j = 0; j < tf_input_tensors[i].shape().dims(); ++j) {
+        ng_shape[j] = tf_input_tensors[i].shape().dim_size(j);
+      }
+      ngraph::element::Type ng_element_type;
+      OP_REQUIRES_OK(ctx,
+          TFDataTypeToNGraphElementType(tf_input_tensors[i].dtype(), &ng_element_type));
+
+      auto backend = BackendManager::GetBackend();
+      std::shared_ptr<ngraph::runtime::Tensor> ng_tensor =
+          backend->create_tensor(ng_element_type, ng_shape, tf_input_tensors[i].data());
+      ng_inputs.push_back(ng_tensor);
+    }
   }
 
   NGRAPH_VLOG(4) << "NGraphEncapsulateOp::Compute allocated argument tensors "
                     "for cluster "
-                 << cluster_id;
+                 << m_cluster_id;
 
   // Allocate tensors for the output results.
 
@@ -280,7 +307,7 @@ void NGraphEncapsulateOp::Compute(OpKernelContext* ctx) {
   }
   NGRAPH_VLOG(4)
       << "NGraphEncapsulateOp::Compute allocated result tensors for cluster "
-      << cluster_id;
+      << m_cluster_id;
 
   int time_create_or_lookup_tensors = create_or_lookup_tensors.ElapsedInMS();
   // Execute the nGraph function.
@@ -291,16 +318,16 @@ void NGraphEncapsulateOp::Compute(OpKernelContext* ctx) {
     {
       NGRAPH_VLOG(4)
           << "NGraphEncapsulateOp::Compute call starting for cluster "
-          << cluster_id;
+          << m_cluster_id;
       try {
         ng_exec->call(ng_inputs, ng_outputs);
       } catch (const std::exception& exp) {
         string status_string = "Caught exception while executing cluster " +
-                               to_string(cluster_id) + string(exp.what());
+                               to_string(m_cluster_id) + string(exp.what());
         OP_REQUIRES(ctx, false, errors::Internal(status_string));
       } catch (...) {
         string status_string =
-            "Caught exception while executing cluster " + to_string(cluster_id);
+            "Caught exception while executing cluster " + to_string(m_cluster_id);
         OP_REQUIRES(ctx, false, errors::Internal(status_string));
       }
     }
@@ -326,19 +353,19 @@ void NGraphEncapsulateOp::Compute(OpKernelContext* ctx) {
   long vm, rss;
   MemoryProfile(vm, rss);
   NGRAPH_VLOG(1) << "NGRAPH_TF_MEM_PROFILE:  OP_ID: "
-                 << ng_encap_impl_.GetInstanceId() << " Step_ID: " << step_id
+                 << m_cluster_id << " Step_ID: " << step_id
                  << " Cluster: " << name() << " Input Tensors created: "
                  << ng_input_tensor_size_in_bytes / (1024 * 1024) << " MB"
                  << " Total process memory: " << rss / (1024 * 1024) << " GB";
 
   NGRAPH_VLOG(4) << "NGraphEncapsulateOp::Compute call done for cluster "
-                 << ng_encap_impl_.GetNgraphCluster();
+                 << m_cluster_id;
 
   NGRAPH_VLOG(4)
       << "NGraphEncapsulateOp::Compute done marking fresh for cluster "
-      << ng_encap_impl_.GetNgraphCluster();
+      << m_cluster_id;
   NGRAPH_VLOG(1) << "NGRAPH_TF_TIMING_PROFILE: OP_ID: "
-                 << ng_encap_impl_.GetInstanceId() << " Step_ID: " << step_id
+                 << m_cluster_id << " Step_ID: " << step_id
                  << " Cluster: " << name()
                  << " Time-Compute: " << compute_time.ElapsedInMS()
                  << " Function-Create-or-Lookup: " << time_func_create_or_lookup
@@ -347,19 +374,109 @@ void NGraphEncapsulateOp::Compute(OpKernelContext* ctx) {
                  << " Execute: " << time_execute_function;
 }  // end compute
 
-int NGraphEncapsulateImpl::s_instance_count = 0;
+// Calls ComputeSignature and gets ngraph executable
+ Status NGraphEncapsulateOp::GetNgExecutable(
+    const std::vector<Tensor>& tf_input_tensors,
+    std::shared_ptr<Executable>& ng_exec) {
+  auto backend = BackendManager::GetBackend();
+
+  // Compute Signature
+  std::vector<TensorShape> input_shapes;
+  std::stringstream signature_ss;
+  for (int i = 0; i < tf_input_tensors.size(); i++) {
+    const Tensor& input_tensor = tf_input_tensors[i];
+    input_shapes.push_back(input_tensor.shape());
+    for (const auto& x : input_tensor.shape()) {
+      signature_ss << x.size << ",";
+    }
+    signature_ss << ";";
+  }
+  signature_ss << "/";
+  for (int i = 0; i < tf_input_tensors.size(); i++) {
+    if (m_input_is_static[i]) {
+      TF_RETURN_IF_ERROR(TensorToStream(signature_ss, tf_input_tensors[i]));
+      signature_ss << ";";
+    }
+  }
+
+  string signature = signature_ss.str();
+  NGRAPH_VLOG(5) << "Computed signature: " << signature;
+  auto it = m_ng_exec_map.find(signature);
+  NGRAPH_VLOG(4) << "NGraphEncapsulateOp::Compute got inputs for cluster "
+                 << m_cluster_id;
+
+  // Translate the TensorFlow graph to nGraph.
+  std::shared_ptr<ngraph::Function> ng_function;
+  if (it == m_ng_exec_map.end()) {
+    // Measure the current total memory usage
+    long vm, rss, vm0, rss0;
+    MemoryProfile(vm0, rss0);
+
+    NGRAPH_VLOG(1) << "Compilation cache miss: " << m_name;
+    std::vector<const Tensor*> static_input_map;
+    for (auto t : tf_input_tensors) {
+      static_input_map.push_back(&t);
+    }
+    TF_RETURN_IF_ERROR(Builder::TranslateGraph(input_shapes, static_input_map,
+                                               &m_graph, ng_function));
+    ng_function->set_friendly_name(m_name);
+
+    if (std::getenv("NGRAPH_TF_DUMP_GRAPHS") != nullptr) {
+      ngraph::plot_graph(ng_function, "tf_function_" + m_name + ".dot");
+    }
+
+    // Evict the cache if the number of elements exceeds the limit
+    std::shared_ptr<Executable> evicted_ng_exec;
+    const char* cache_depth_specified =
+        std::getenv("NGRAPH_TF_FUNCTION_CACHE_ITEM_DEPTH");
+    if (cache_depth_specified != nullptr) {
+      m_function_cache_depth_in_items = atoi(cache_depth_specified);
+    }
+    if (m_ng_exec_map.size() >= m_function_cache_depth_in_items) {
+      evicted_ng_exec = m_ng_exec_map[m_lru.back()];
+      m_ng_exec_map.erase(m_lru.back());
+
+      // Call delete function here for the erased func
+      backend->remove_compiled_function(evicted_ng_exec);
+
+      m_lru.pop_back();
+    }  // cache eviction if cache size greater than cache depth
+
+    NG_TRACE("Compile nGraph", m_name, "");
+    try {
+      ng_exec = backend->compile(ng_function);
+    } catch (const std::exception& ex) {
+      return errors::Internal("Failed to compile function " + m_name + ": ",
+                              ex.what());
+    }
+
+    m_ng_exec_map[signature] = ng_exec;
+    m_lru.push_front(signature);
+
+    // Memory after
+    MemoryProfile(vm, rss);
+    auto delta_vm_mem = vm - vm0;
+    auto delta_res_mem = rss - rss0;
+    NGRAPH_VLOG(1) << "NGRAPH_TF_CACHE_PROFILE: OP_ID: " << m_cluster_id
+                   << " Cache length: " << m_ng_exec_map.size()
+                   << " Cluster: " << m_name << " Delta VM: " << delta_vm_mem
+                   << " Delta RSS: " << delta_res_mem
+                   << " KB Total RSS: " << rss / (1024 * 1024) << " GB "
+                   << " VM: " << vm / (1024 * 1024) << " GB" << endl;
+  }  // end of input signature not found in m_ng_exec_map
+  else {
+    // Found the input signature in m_ng_exec_map, use the cached executable
+    // Update the m_lru
+    if (signature != m_lru.front()) {
+      m_lru.remove(signature);
+      m_lru.push_front(signature);
+    }
+    ng_exec = it->second;
+  }
+  return Status::OK();
+}
 
 }  // namespace ngraph_bridge
-
-REGISTER_OP("_nGraphEncapsulate")
-    .Input("args: Targuments")
-    .Attr("Targuments: list(type) >= 0")
-    .Output("results: Tresults")
-    .Attr("Tresults: list(type) >= 0")
-    .Attr("ngraph_cluster: int")
-    .Attr("ngraph_graph_id: int")
-    .SetIsStateful()
-    .Doc("nGraph Encapsulation Op. For use by the nGraph JIT only.");
 
 REGISTER_KERNEL_BUILDER(Name("_nGraphEncapsulate").Device(DEVICE_CPU),
                         ngraph_bridge::NGraphEncapsulateOp);
