+++ ngraph_bridge/ops/ngraph_encapsulate_op.cc
@@ -1,0 +1,33 @@
+/*******************************************************************************
+ * Copyright 2017-2020 Intel Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *******************************************************************************/
+
+#include "tensorflow/core/common_runtime/optimization_registry.h"	
+
+namespace tensorflow {	
+namespace ngraph_bridge {	
+
+REGISTER_OP("_nGraphEncapsulate")
+    .Input("args: Targuments")
+    .Attr("Targuments: list(type) >= 0")
+    .Output("results: Tresults")
+    .Attr("Tresults: list(type) >= 0")
+    .Attr("ngraph_cluster: int")
+    .Attr("ngraph_graph_id: int")
+    .SetIsStateful()
+    .Doc("nGraph Encapsulation Op. For use by the nGraph JIT only.");
+
+}  // namespace ngraph_bridge	
+}  // namespace tensorflow
